requirements.txt
streamlit
geopandas
matplotlib.pyplot 
os
folium
requests
pandas